refactor: apply PascalCase for react components
