+++ components/library/widget/code-completion-widget/Component.tsx
