+++ components/library/widget/quiz-widget/Component.tsx
