+++ eslint.config.mjs
@@ -64,6 +64,15 @@ const eslintConfig = defineConfig([
       'unicorn/no-array-callback-reference': 'off',
       'unicorn/no-array-for-each': 'off',
       'unicorn/no-array-reduce': 'off',
+      'unicorn/filename-case': [
+        'error',
+        {
+          cases: {
+            kebabCase: true,
+            pascalCase: true,
+          },
+        },
+      ],
 
       'simple-import-sort/imports': 'error',
       'simple-import-sort/exports': 'error',
